requirements.txt
# ============================================================
# Benamdaj Hybrid RAG System — Requirements
# BDS-RAG-001 | Patricia Orji
# MSc AI & Data Science, University of Wolverhampton
#
# Install: pip install -r requirements.txt
# ============================================================

# ── Core NLP & Retrieval ─────────────────────────────────────
rank-bm25==0.2.2
sentence-transformers>=2.7,<3.0
nltk>=3.8

# ── Vector database ──────────────────────────────────────────
pinecone>=4.0,<6.0

# ── LLM / OpenAI ─────────────────────────────────────────────
openai>=1.30,<2.0

# ── Document parsing ─────────────────────────────────────────
pdfplumber>=0.10

# ── Orchestration (optional) ──────────────────────────────────
langchain>=0.2,<0.4
langchain-openai>=0.1,<0.3
langchain-community>=0.2,<0.4

# ── Evaluation ───────────────────────────────────────────────
ragas>=0.1,<0.3
datasets>=2.18

# ── Web server ───────────────────────────────────────────────
Flask>=3.0

# ── Data & utilities ──────────────────────────────────────────
numpy>=1.26
pandas>=2.2
python-dotenv>=1.0
tqdm>=4.66

# ── Testing ───────────────────────────────────────────────────
pytest>=8.0
